+++ packages/react/src/components/header/header.react-template.tsx
@@ -17,6 +17,8 @@ export const reactHeader: ComponentImplementation<Header> = {
       authStatus,
       loginUrl,
       logoutUrl,
+      showHelp,
+      helpUrl,
       userProfileUrl,
       userProfileName,
       userHomeUrl,
@@ -30,6 +32,8 @@ export const reactHeader: ComponentImplementation<Header> = {
           authStatus={authStatus}
           loginUrl={loginUrl}
           logoutUrl={logoutUrl}
+          showHelp={showHelp}
+          helpUrl={helpUrl}
           userProfileUrl={userProfileUrl}
           userProfileName={userProfileName}
           userHomeUrl={userHomeUrl}