+++ storybook/src/components/header/header.core-template.ts
@@ -15,6 +15,8 @@ export const coreHeader: ComponentImplementation<Header> = {
       authStatus,
       loginUrl,
       logoutUrl,
+      showHelp,
+      helpUrl,
       userProfileUrl,
       userProfileName,
       userHomeUrl,
@@ -27,6 +29,8 @@ export const coreHeader: ComponentImplementation<Header> = {
         auth-status=${ifDefined(authStatus)}
         login-url=${ifDefined(loginUrl)}
         logout-url=${ifDefined(logoutUrl)}
+        ?show-help=${showHelp}
+        help-url=${ifDefined(helpUrl)}
         user-profile-url=${ifDefined(userProfileUrl)}
         user-profile-name=${ifDefined(userProfileName)}
         user-home-url=${ifDefined(userHomeUrl)}